requirements.txt
appnope==0.1.0
backcall==0.2.0
boto==2.49.0
boto3==1.9.191
botocore==1.12.253
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
decorator==4.4.2
dnspython==1.16.0
docutils==0.15.2
Flask==1.1.1
gunicorn==20.0.4
idna==2.9
ipython==7.15.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
lockfile==0.12.2
luigi==3.0.0
lxml==4.3.2
MarkupSafe==1.1.1
numpy==1.16.2
pandas==0.24.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pre-commit==2.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pymongo==3.10.1
python-daemon==2.2.4
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
s3transfer==0.2.1
selenium==3.141.0
six==1.15.0
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.3
tzlocal==2.0.0
urllib3==1.25.9
wcwidth==0.2.4
Werkzeug==1.0.1
wget==3.2
yapf==0.30.0
